cleanup and eslint
