+++ .eslintrc.cjs
@@ -12,6 +12,7 @@ module.exports = {
   settings: { react: { version: '18.2' } },
   plugins: ['react-refresh'],
   rules: {
+    "react/prop-types": "off",
     'react/jsx-no-target-blank': 'off',
     'react-refresh/only-export-components': [
       'warn',
@@ -19,3 +20,4 @@ module.exports = {
     ],
   },
 }
+
